+++ src/py/app/domain/tags/schemas/_tag.py
@@ -1,5 +1,6 @@
 """Tag schemas."""
 
+from datetime import datetime
 from uuid import UUID
 
 from app.lib.schema import CamelizedBaseStruct
@@ -11,6 +12,8 @@ class Tag(CamelizedBaseStruct):
     id: UUID
     slug: str
     name: str
+    created_at: datetime
+    updated_at: datetime
 
 
 class TagCreate(CamelizedBaseStruct):